+++ packages/soundxyz/src/index.ts
@@ -1,12 +1,16 @@
 import {
   type IActionPlugin,
   PluginActionNotImplementedError,
-} from '@rabbitholegg/questdk'
+  type ActionParams,
+  type MintActionParams,
+} from '@rabbitholegg/questdk-plugin-utils'
 
 import {
   mint,
   getSupportedChainIds,
   getSupportedTokenAddresses,
+  getMintIntent,
+  getProjectFees,
 } from './Soundxyz.js'
 
 export const Soundxyz: IActionPlugin = {
@@ -16,4 +20,7 @@ export const Soundxyz: IActionPlugin = {
   mint,
   bridge: async () => new PluginActionNotImplementedError(),
   swap: async () => new PluginActionNotImplementedError(),
+  getMintIntent,
+  getProjectFees: async (params: ActionParams) =>
+    getProjectFees(params as unknown as MintActionParams),
 }